Expand checks for bad clothing when trying to put on an outfit. Also let the player know why the clothing cannot be worn if it is bad.
The wardrobe outfits had a weird bug when trying to wear clothing that the player was not allowed to anymore.

The previous code left the player wearing clothing = 0 if they put on an outfit with clothing that they could not wear. Clothing = 0 does not seem to be defined since it had no image or attributes but does not count as naked either so it fell through the other clothing checks in the game for nudity and skirts and probably others too.

I kept the if statements the same when checking for bad clothing but moved them and expanded them so that the player knows why the clothing cannot be worn so they can try and remedy the situation. Also the player will be correctly left not wearing clothing.

I feel like these new checks could benefit from images of generic dressing/undressing but I don't know what the rules are for images so it is just text right now. New code is on my fork.